add maxId
right so this is to do with my concept of "anchoring." For sort of complicated reasons, I think I want to stop rule34 from adding new posts if I want my caching to work. So I'm planning to include `id:<${maxId}` in every prompt, so that no new posts past a certain point (that point being maxId) are allowed in. Every once in a while I think the user can manually reset the anchor, but it's not a big deal I feel like.