get rid of SearchCache and PostsCache and add PercentCache and PercentTags()
I got rid of postcache and searchcache because I really felt like it was too big to be effective. saving and loading the cache was just mad slow, and it was like "I don't think this is gonna work." But then I was thinking, and I was like "if I only cached the answers to questions like 'how many tagA are in feet posts?' that would hardly take up any space at all!" So that's where this all came from.