fix: stop using rawPosts
I was for some reason thinking it would be good to use rawposts bc then I could save time having to process them unnecarily. But I forgot rawposts don't even have their tags in arrays, and that was causing my code to not work. So I switched back to using processed posts and it works now